Remove using namespace std and most size_t

+++ phi_block.hpp
@@ -28,8 +28,8 @@ inline int64_t ceil_div(int64_t x, int64_t y)
 class PhiBlock
 {
 public:
-    size_t           zk1;       // z_{k-1}, block lower bound (inclusive)
-    size_t           zk;        // z_k, block upper bound (exclusive)
+    int64_t          zk1;       // z_{k-1}, block lower bound (inclusive)
+    int64_t          zk;        // z_k, block upper bound (exclusive)
     size_t           bsize;     // logical block size
     size_t           psize;     // physical block size
  
@@ -48,14 +48,14 @@ public:
     }
 
     // translate into actual index into tree
-    size_t tree_index(const size_t y) const
+    size_t tree_index(const int64_t y) const
     {
         return (y - zk1)/2;
     }
 
     // sum contribution of this block to phi(y,b)
     // = phi(y,b) - phi(zk1 - 1, b) base
-    int64_t sum_to(size_t y) const
+    int64_t sum_to(int64_t y) const
     {
         assert(y >= zk1);
         assert(y < zk);
@@ -67,16 +67,13 @@ public:
     {
         assert(pb > 2); // 2 already sieved by default
 
-        size_t jstart = pb * ceil_div(zk1, pb);
+        int64_t jstart = pb * ceil_div(zk1, pb);
         if (jstart % 2 == 0)
             jstart += pb; // ensure odd
 
-        for (size_t j = jstart; j < zk; j += 2*pb)
+        for (int64_t j = jstart; j < zk; j += 2*pb)
         {
             phi_sum.try_decrease(tree_index(j));
         }
     }
-
 };
-
-
